requirements.txt
numpy
scipy
networkx
pytorch
matplotlib
pandas